server/pyproject.toml
[project]
name = "reranker-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-health-checking>=1.71.2",
    "grpcio-tools==1.68.1",
    "numpy>=2.3.5",
    "onnxruntime>=1.23.2",
    "protobuf==5.29.5",
    "psutil>=7.1.3",
    "threadpoolctl>=3.6.0",
    "tokenizers>=0.21.4",
]

[tool.uv.workspace]
members = [
    "model",
]
